# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rest Api",
        "contact": {
            "name": "Fúlvio Cezar Canducci Dias",
            "email": "fulviocanducci@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "App\\Http\\Controllers\\LoginController::index",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login User"
                    }
                }
            }
        },
        "/api/people": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "List Of People",
                "description": "List Of People",
                "operationId": "getProjectsList",
                "responses": {
                    "200": {
                        "description": "List of People",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/peoples"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "People"
                ],
                "summary": "Create People",
                "description": "Create People",
                "operationId": "App\\Http\\Controllers\\PeopleController::create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/people"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create People",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/people"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "People"
                ],
                "summary": "Show People By Id",
                "description": "Show People By Id",
                "operationId": "App\\Http\\Controllers\\PeopleController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show People By Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/people"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "People"
                ],
                "summary": "Update People",
                "description": "Update People",
                "operationId": "App\\Http\\Controllers\\PeopleController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/people"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update People",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/people"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "People"
                ],
                "summary": "Delete People By Id",
                "description": "Delete People By Id",
                "operationId": "App\\Http\\Controllers\\PeopleController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Delete People By Id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/destroy"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "destroy": {
                "properties": {
                    "found": {
                        "type": "boolean"
                    },
                    "deleted": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "login": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "people": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "peoples": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Project Example ApiRest",
            "description": "ApiRest"
        }
    ],
    "security": [
        []
    ]
}